refactored db access operations and tested that we can still connect
