+++ server/modules/dependencies/db_session.py
@@ -1,0 +1,11 @@
+from ..pool import SessionLocal
+
+# a dependency for routes
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
