+++ server/modules/pool.py
@@ -10,60 +10,60 @@ from pydantic import BaseModel
 import os
 from dotenv import load_dotenv
 
+''' This is all a bunch of set up for the database - could rename the file db_config.py'''
+
 load_dotenv(os.path.dirname(os.path.dirname(os.getcwd()))+'/.env')
 
-
-# SQLALCHEMY_DATABASE_URL = "postgresql://user:password@pg_server_name/database_name"
 SQLALCHEMY_DATABASE_URL = os.environ.get('DATABASE_URL') or "postgresql+psycopg2://dan:password@localhost:5432/prime_app"
 
 engine = create_engine(
     SQLALCHEMY_DATABASE_URL
 )
+
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 Base = declarative_base()
-
-class User(Base):
-    __tablename__ = "user"
-
-    id = Column(Integer, primary_key=True, index=True)
-    username = Column(String, unique=True, index=True)
-    password = Column(String) 
-
-class UserBase(BaseModel):
-    username: str
-
-    class Config:
-        orm_mode = True
-    
-class UserCreate(UserBase):
-    password: str
-
-def get_user(db: Session, user_id: int):
-    return db.query(User).filter(User.id == user_id).first()
-
-def create_user(db: Session, user: UserCreate):
-    fake_password = user.password
-    db_user = User(username=user.username, password=fake_password)
-    db.add(db_user)
-    db.commit()
-    db.refresh(db_user)
-    return db_user
-
 Base.metadata.create_all(bind=engine)
 
-# a dependency for routes
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
+# class User(Base):
+#     __tablename__ = "user"
+#     id = Column(Integer, primary_key=True, index=True)
+#     username = Column(String, unique=True, index=True)
+#     password = Column(String) 
 
-# purely for testing
-with SessionLocal() as db:
-    try:
-        create_user(db, User(username='marko6', password='123abc'))
-        new_user = get_user(db, 7)
-    finally:
-        db.close()
+# class UserBase(BaseModel):
+#     username: str
+
+#     class Config:
+#         orm_mode = True
+    
+# class UserCreate(UserBase):
+#     password: str
+
+# def get_user(db: Session, user_id: int):
+#     return db.query(User).filter(User.id == user_id).first()
+
+# def create_user(db: Session, user: UserCreate):
+#     fake_password = user.password
+#     db_user = User(username=user.username, password=fake_password)
+#     db.add(db_user)
+#     db.commit()
+#     db.refresh(db_user)
+#     return db_user
+
+
+# # a dependency for routes
+# def get_db():
+#     db = SessionLocal()
+#     try:
+#         yield db
+#     finally:
+#         db.close()
+
+# # purely for testing
+# with SessionLocal() as db:
+#     try:
+#         create_user(db, User(username='marko6', password='123abc'))
+#         new_user = get_user(db, 7)
+#     finally:
+#         db.close()
